// frontend/src/types/statistici.ts
export interface PasariPeMutatie {
  mutatie: string;
  total: number;
}

export interface PuiPeAn {
  an: number;
  total: number;
}

export interface ProductiePereche {
  perecheId: string;
  mascul: string;
  femela: string;
  total: number;
}

export interface ProductieParinte {
  id: string;
  nrInel: string;
  total: number;
}